+++ docs/reference/models/classifiers/wav2vec2_classifier.md
@@ -1,3 +1,3 @@
 # Wav2Vec2 Classifier
 
-::: speechline.ml.classifiers.Wav2Vec2Classifier
+::: speechline.classifiers.Wav2Vec2Classifier
