+++ docs/reference/models/transcribers/wav2vec2_transcriber.md
@@ -1,3 +1,3 @@
 # Wav2Vec2 Transcriber
 
-::: speechline.ml.transcribers.Wav2Vec2Transcriber
+::: speechline.transcribers.Wav2Vec2Transcriber
