+++ docs/reference/models/transcribers/whisper_transcriber.md
@@ -1,3 +1,3 @@
 # Whisper Transcriber
 
-::: speechline.ml.transcribers.WhisperTranscriber
+::: speechline.transcribers.WhisperTranscriber
