+++ docs/reference/pipelines/audio_classification_with_padding_pipeline.md
@@ -1,3 +1,3 @@
 # Audio Classification with Padding Pipeline
 
-::: speechline.ml.pipelines.AudioClassificationWithPaddingPipeline
+::: speechline.pipelines.AudioClassificationWithPaddingPipeline
